pyproject.toml
[tool.poetry]
name = "agileupstate"
version = "0.0.41"
description = "State mangement tool for Future Agile CICD"
authors = ["agileturret <Paul.Gilligan@agilesolutions.co.uk>"]
license = "Apache-2.0"
readme = "README.md"
homepage = "https://github.com/Agile-Solutions-GB-Ltd/agileupstate"
repository = "https://github.com/Agile-Solutions-GB-Ltd/agileupstate.git"

[tool.poetry.dependencies]
python = "^3.8"
click = "^8.1"
PyYAML = "^5.4.1"
prettytable = "^3.3.0"
hvac = "^1.0.2"
pywinrm = "^0.4.3"

[tool.poetry.dev-dependencies]
flake8 = "^4.0"
bumpversion = "^0.6"
pytest = "^7.1.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry_bumpversion.file."src/agileupstate/__init__.py"]

[tool.poetry.scripts]
agileupstate = "agileupstate.cli:cli"
